server/requirements.txt
alembic==1.16.4 ; python_version >= "3.13" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.13" and python_version < "4.0"
anyio==4.10.0 ; python_version >= "3.13" and python_version < "4.0"
apscheduler==3.11.0 ; python_version >= "3.13" and python_version < "4.0"
babel==2.17.0 ; python_version >= "3.13" and python_version < "4.0"
certifi==2025.8.3 ; python_version >= "3.13" and python_version < "4.0"
charset-normalizer==3.4.3 ; python_version >= "3.13" and python_version < "4.0"
click==8.2.1 ; python_version >= "3.13" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.13" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
courlan==1.3.2 ; python_version >= "3.13" and python_version < "4.0"
dateparser==1.2.2 ; python_version >= "3.13" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.13" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.13" and python_version < "4.0"
fastapi-cli[standard]==0.0.8 ; python_version >= "3.13" and python_version < "4.0"
fastapi-cloud-cli==0.1.5 ; python_version >= "3.13" and python_version < "4.0"
fastapi[standard]==0.116.1 ; python_version >= "3.13" and python_version < "4.0"
feedparser==6.0.11 ; python_version >= "3.13" and python_version < "4.0"
greenlet==3.2.4 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.13"
h11==0.16.0 ; python_version >= "3.13" and python_version < "4.0"
htmldate==1.9.3 ; python_version >= "3.13" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.13" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.13" and python_version < "4.0"
httpx==0.27.2 ; python_version >= "3.13" and python_version < "4.0"
idna==3.10 ; python_version >= "3.13" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.13" and python_version < "4.0"
justext==3.0.2 ; python_version >= "3.13" and python_version < "4.0"
lxml-html-clean==0.4.2 ; python_version >= "3.13" and python_version < "4.0"
lxml==5.4.0 ; python_version >= "3.13" and python_version < "4.0"
lxml[html-clean]==5.4.0 ; python_version >= "3.13" and python_version < "4.0"
mako==1.3.10 ; python_version >= "3.13" and python_version < "4.0"
markdown-it-py==4.0.0 ; python_version >= "3.13" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.13" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.13" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.13" and python_version < "4.0"
pydantic==2.11.7 ; python_version >= "3.13" and python_version < "4.0"
pydantic[email]==2.11.7 ; python_version >= "3.13" and python_version < "4.0"
pygments==2.19.2 ; python_version >= "3.13" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.13" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.13" and python_version < "4.0"
python-multipart==0.0.20 ; python_version >= "3.13" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.13" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.13" and python_version < "4.0"
regex==2025.7.34 ; python_version >= "3.13" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.13" and python_version < "4.0"
rich-toolkit==0.15.0 ; python_version >= "3.13" and python_version < "4.0"
rich==14.1.0 ; python_version >= "3.13" and python_version < "4.0"
rignore==0.6.4 ; python_version >= "3.13" and python_version < "4.0"
sentry-sdk==2.35.0 ; python_version >= "3.13" and python_version < "4.0"
sgmllib3k==1.0.0 ; python_version >= "3.13" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.13" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.13" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.13" and python_version < "4.0"
sqlalchemy==2.0.43 ; python_version >= "3.13" and python_version < "4.0"
starlette==0.47.2 ; python_version >= "3.13" and python_version < "4.0"
tld==0.13.1 ; python_version >= "3.13" and python_version < "4.0"
trafilatura==2.0.0 ; python_version >= "3.13" and python_version < "4.0"
typer==0.16.0 ; python_version >= "3.13" and python_version < "4.0"
typing-extensions==4.14.1 ; python_version >= "3.13" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.13" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.13" and python_version < "4.0" and platform_system == "Windows"
tzlocal==5.3.1 ; python_version >= "3.13" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.13" and python_version < "4.0"
uvicorn[standard]==0.35.0 ; python_version >= "3.13" and python_version < "4.0"
uvloop==0.21.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.13" and python_version < "4.0"
watchfiles==1.1.0 ; python_version >= "3.13" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.13" and python_version < "4.0"
